+++ src/parser/api.ts
@@ -1,3 +1,4 @@
+export { header } from './api/header';
 export { parse } from './api/parse';
 export { bind } from './api/bind';
 export { caches } from './api/cache';