+++ src/parser/api/header.test.ts
@@ -1,0 +1,19 @@
+import { header } from './header';
+
+describe('Unit: parser/api/header', () => {
+  describe('header', () => {
+    it('', () => {
+      assert.deepStrictEqual(header(''), undefined);
+      assert.deepStrictEqual(header('---'), undefined);
+      assert.deepStrictEqual(header('---\n'), undefined);
+      assert.deepStrictEqual(header('---\n---'), []);
+      assert.deepStrictEqual(header('---\n---\n'), []);
+      assert.deepStrictEqual(header('---\na: b\n'), undefined);
+      assert.deepStrictEqual(header('---\na: b\n---'), ['a: b']);
+      assert.deepStrictEqual(header('---\na: b\nc: d\n---'), ['a: b', 'c: d']);
+      assert.deepStrictEqual(header('---\na: b\nc: d\n---\n'), ['a: b', 'c: d']);
+    });
+
+  });
+
+});